model/requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
scikit-learn==1.5.1
numpy==1.26.4
pydantic==2.8.2
opencv-python==4.10.0.84
tensorflow==2.17.0
keras==3.5.0
pillow==10.2.0
matplotlib==3.9.2
torch==2.4.1
torchaudio==2.4.1
yolov5==7.0.13